refactor(ui): use redux-provided hooks for accessing store
