+++ invokeai/frontend/web/src/app/store/use-debounced-app-selector.ts
@@ -1,5 +1,5 @@
 import type { Selector } from '@reduxjs/toolkit';
-import { useAppStore } from 'app/store/nanostores/store';
+import { useAppStore } from 'app/store/storeHooks';
 import type { RootState } from 'app/store/store';
 import { useEffect, useState } from 'react';
