+++ invokeai/frontend/web/src/features/controlLayers/components/RefImage/RefImageList.tsx
@@ -1,5 +1,5 @@
 import { Button, Collapse, Divider, Flex } from '@invoke-ai/ui-library';
-import { useAppStore } from 'app/store/nanostores/store';
+import { useAppStore } from 'app/store/storeHooks';
 import { useAppSelector } from 'app/store/storeHooks';
 import { useImageUploadButton } from 'common/hooks/useImageUploadButton';
 import { RefImagePreview } from 'features/controlLayers/components/RefImage/RefImagePreview';
