+++ invokeai/frontend/web/src/features/dynamicPrompts/hooks/useDynamicPromptsWatcher.tsx
@@ -1,4 +1,4 @@
-import { useAppStore } from 'app/store/nanostores/store';
+import { useAppStore } from 'app/store/storeHooks';
 import { useAppSelector } from 'app/store/storeHooks';
 import { debounce } from 'es-toolkit/compat';
 import {
