+++ invokeai/frontend/web/src/features/gallery/components/ImageGrid/GalleryImage.tsx
@@ -3,7 +3,7 @@ import { draggable, monitorForElements } from '@atlaskit/pragmatic-drag-and-drop
 import type { SystemStyleObject } from '@invoke-ai/ui-library';
 import { Box, Flex, Icon, Image } from '@invoke-ai/ui-library';
 import { createSelector } from '@reduxjs/toolkit';
-import { useAppStore } from 'app/store/nanostores/store';
+import { useAppStore } from 'app/store/storeHooks';
 import type { AppDispatch, AppGetState } from 'app/store/store';
 import { useAppSelector } from 'app/store/storeHooks';
 import { uniq } from 'es-toolkit';
